+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- A builtin extension for `telescope` - see `:h compile-mode-telescope-extension`
+
 ## [5.12.0] - 2026-01-02
 
 Also includes several documentation changes.